requirements.txt
openai>=1.66.0
python-dotenv>=1.0.1
PyMuPDF>=1.25.0
python-docx>=1.1.2
qdrant-client>=1.13.2
pandas>=2.2.3
openpyxl>=3.1.5
langchain-text-splitters>=0.3.6
